Improved mystery switch and tablet media queries

+++ js/main.js
@@ -699,18 +699,40 @@ function HomeJS() {
       var $beadSelected = $('button.bead--selected');
 
       if ($currPrayer.attr('id') === 'prayer-4' && $beadSelected.attr('id') === 'medallion') {
+        var byDayReminder = '';
+
+        if (!localStorage.getItem('mysteries')) {
+          byDayReminder = ' You will need to switch back to "By Day" in Settings tomorrow.';
+        }
+
         if (_store__WEBPACK_IMPORTED_MODULE_0__["default"].settings.mysteries === 'joyful') {
-          _store__WEBPACK_IMPORTED_MODULE_0__["default"].settings.mysteries = 'luminous';
-          localStorage.setItem('mysteries', 'luminous');
+          var advanceMysteries = confirm('Would you like to switch to the Luminous mysteries?' + byDayReminder);
+
+          if (advanceMysteries) {
+            _store__WEBPACK_IMPORTED_MODULE_0__["default"].settings.mysteries = 'luminous';
+            localStorage.setItem('mysteries', 'luminous');
+          }
         } else if (_store__WEBPACK_IMPORTED_MODULE_0__["default"].settings.mysteries === 'luminous') {
-          _store__WEBPACK_IMPORTED_MODULE_0__["default"].settings.mysteries = 'sorrowful';
-          localStorage.setItem('mysteries', 'sorrowful');
+          var advanceMysteries = confirm('Would you like to switch to the Sorrowful mysteries?' + byDayReminder);
+
+          if (advanceMysteries) {
+            _store__WEBPACK_IMPORTED_MODULE_0__["default"].settings.mysteries = 'sorrowful';
+            localStorage.setItem('mysteries', 'sorrowful');
+          }
         } else if (_store__WEBPACK_IMPORTED_MODULE_0__["default"].settings.mysteries === 'sorrowful') {
-          _store__WEBPACK_IMPORTED_MODULE_0__["default"].settings.mysteries = 'glorious';
-          localStorage.setItem('mysteries', 'glorious');
+          var advanceMysteries = confirm('Would you like to switch to the Glorious mysteries?' + byDayReminder);
+
+          if (advanceMysteries) {
+            _store__WEBPACK_IMPORTED_MODULE_0__["default"].settings.mysteries = 'glorious';
+            localStorage.setItem('mysteries', 'glorious');
+          }
         } else if (_store__WEBPACK_IMPORTED_MODULE_0__["default"].settings.mysteries === 'glorious') {
-          _store__WEBPACK_IMPORTED_MODULE_0__["default"].settings.mysteries = 'joyful';
-          localStorage.setItem('mysteries', 'joyful');
+          var advanceMysteries = confirm('Would you like to switch to the Joyful mysteries?' + byDayReminder);
+
+          if (advanceMysteries) {
+            _store__WEBPACK_IMPORTED_MODULE_0__["default"].settings.mysteries = 'joyful';
+            localStorage.setItem('mysteries', 'joyful');
+          }
         }
 
         $('#intro-our-father-2').click();
